todo_backend/requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
Pillow==10.0.1
python-decouple==3.8
PyJWT==2.10.1
sqlparse==0.5.3
